DeviceMotionEvent: Add a note about acceleration properties.
See #96 for an example of how these attributes can be confusing, especially
given accelerationIncludingGravity's name.

Add a note with a reference to what exactly that property measures, and
refence the Motion Sensors Explainer. It contains a more detailed
explanation about Accelerometer and LinearAccelerationSensor, which is
exactly the difference between the accelerationIncludingGravity and
acceleration attributes.