editorial: Move explanations about orientation and motion to a new "Model" section
The idea is to have the more general discussion about what is being measured
and how in one section separate from the section that covers the API, each
event's attributes and security requirements.

The device orientation contents in particular were quite long and mixed with
normative IDL descriptions and requirements that made the text confusing to
follow.

The new "device orientation model" section has separated the description of
the rotation steps from the discussion of different reference frames.

The new "device motion model" section has some of the old content, but more
text has been added along with references to the MOTION-SENSORS,
ACCELEROMETER and GYROSCOPE specs which not only implement the same
functionality but also have more in-depth content about the same topics.