Add Permissions Policy integration
This substantive change makes the firing of the events defined in this
specification dependent on a few different policy-controlled features:
"accelerometer", "gyroscope", and "magnetometer.

Both Blink and WebKit already integrate their implementations with the
Permissions Policy spec and use the same tokens, although a bit differently:

- Both require "accelerometer" and "gyroscope" for the devicemotion event.
- Blink requires "accelerometer" and "gyroscope" to provide relative
  orientation data for the deviceorientation event, and additionally the
  "magnetometer" feature to fall back to absolute orientation data. WebKit
  always requires the three tokens for the deviceorientation event.
- Blink requires "accelerometer", "gyroscope", and "magnetometer" for the
  deviceorientationabsolute event, whereas WebKit does not implement this
  event.

We have opted to codify Blink's behavior in the specification, as since
relative orientation data does not require a magnetometer sensor, the event
should not require the corresponding token either.

Fixes #64.