Make DeviceMotionEventInit members nullable with null default
This change reverts #55 and defines the default values of the
DeviceMotionEventInit as null as I believe was the original intent of
the definition before that change.

Null DeviceMotionEvent attributes have semantic meaning (they declare
that the host does not provide the given sensor) and so script should
be able to initialize an event with its attributes set to null.

Fixed #91.